trainer py files are renamed
